Use iv_generate_video_uri() in iviewiir
Makes changes to the iviewiir_series and iviewiir_download interfaces.
iviewiir_series now takes a struct iv_item ** as a parameter which
when it returns will contain the series list. iviewiir_download takes in
a struct iv_item *, which is the item to whose associated video should
be downloaded.

Also fixes setting of the prefix member of the struct iv_auth instance
in iv_parse_auth.