Use the auth xml URI from the config XML
Previously we were using a hardcoded value.  This broke
downloading of shows when the ABC changed their authentication
url recently.  It turns out the correct value was being
parsed and stored, but not used.  Now we use the parsed value,
and therefore the hard coded value can be removed.

Thanks to Jeremy for picking up this, and providing example code in python-iview:

  https://jeremy.visser.name/2010/11/10/python-iview-update-authentication-fix/