Annotate unused function arguments
The SAX parser code has a whole heap of callback functions where we must use
the given method signature, leading to lots of unused variables.  This patch
annotates them with __attribute__((unused)) in order to silence warnings
produced at verbose error levels.

This patch is the first step on the way to enabling -Wall -Werror.