Replace optparse with CCAN's opt module
Inspired by Rusty's talk at linux.conf.au 2011, I used the opt module from CCAN
to replace our usage of optparse.  The code is more readable, and there are
fewer lines in iviewiir.c, however we are now carrying around three CCAN
modules.

The licence information for each module is in the unmodified _info file in each
directory.  The test directory has been removed in each case, as has the
LICENCE symlink, but aside from this the module directories are an exact copy
of CCAN as of 59d2b90f59df57007c97cd8f8dc3db9e24781e06.

The usage message has lost a lot of text that will be re-added in the form of a
manpage:

    $ iviewiir -h
    Usage: iviewiir [-isafh] [SID[:PID]]
    --items-list|-i <arg> List episodes in a series. Requires a SID as a parameter.
    --series-list|-s    List the series available. The first element is the SID.
    --all|-a            List all items in all non-empty series.
    --force|-f          Force bypass the cached metadata.
    --help|-h           Show this message.

The relevant c files have been added to the IVIEWIIR target in the Makefile,
and a include path of the top level directory was added in order to find the
headers, as CCAN modules's include path is always relative to the top level.