Unbreak options by moving test
As ccan/opt decrements argc when an option has been parsed, argc will equal 1
when printing the program, causing the help message to be displayed and exit.
Instead test for argc == 1 after the options have been evaluated.

Also, fix up comments to use c commenting style to match the rest of the file.