Use the SAX2 parser from libxml2
It turns out that libxml2 has an old and a new SAX parser.  This only affects
the internal API, as we can use the same function calls in the same way.  The
only time you will notice is trying to link with a libxml2 that has been
configured with --without-sax1.

This patch sets the magic member variable ->initialized to XML_SAX2_MAGIC, thus
freeing the link time dependency on SAX1.