iviewiir: Fix off-by-one error when printing series items
An older implementation of the items XML parsing used to trigger on the
<title> tag describing the series, not an actual episode. As a
workaround, iviewiir would skip the first entry in the items list to
avoid printing information that wasn't about a specific episode. With a
later implementation of the XML parser this issue disappeared, but
iviewiir was never changed to suit the new behaviour. Thanks to Adam
Hills for helping discover this bug. We were trying to watch your new
show.