wiicode/http: Support limited chunked encoding
When HTTP 1.1 sends the header Transfer-Encoding: chunked, the client must
detect this in the header, and then decode the payload as a series of chunks.
Each chunk begins with a hex number that indicates the number of octets in that
chunk.

The previous http code had no support for this, and would fail with a strange
combination of errors and success indicators.  This adds support for decoding a
HTTP transfer with a single variable length chunk, which is sufficient for our
needs.