Bail out earlier when no arguments are presented
With ccan/opt, printing the help message has to happen after parsing the
options (or else it segfaults when calling usage_and_exit). As parsing the
options reduces argc, we can't use that anymore for testing if nothing was
passed. So we have to test each of the options individually.  This does save us
from hitting the network when no arguments are passed.