libiview/Makefile: Honour CC variable in libiview
Since v0.2-55-g44ecfc6, where support for using a cross compiler was added,
Libiview hard-coded the c compiler to gcc in the libiview makefile.  This
breaks using llvm and other alternate CC lines.

Note that the := is important to ensure that the assignment is not considered a
recursive definition by make.